Simply implementation of rootd / client communication
Instead of spawning four pipes (CMD, stdout, stderr, RC) on top of the
existing UDS communication layer, just move everything through the UDS.

Client sends a command as a message (up to 4096 bytes), along with an
ancillary cmsg to pass the file descriptors for its stdin, stdout, and
stderr. The rootd will validate the command etc. as before, and simply
execute the command with those file descriptors attached, and send the
return code as its response over the UDS.

The client is expected to run the fake internal "af-init" command after
first opening the connection to the UDS. This passes the CDIR to rootd
and lets it know if bootstrapping is required or not.

The UDS connection file descriptor is inheritable and can be reused for
multiple commands so long as "af-init" was run as the first command.