objects: fixup "created" property for MQTT
This will be important once artifact synchronization support is added.
The plan is to tie the timestamps of at least the REPODEST symlinks to
the creation timestamp of the relevant job, thus ensuring that the
latest job always has the right-of-way to update those symlinks if using
e.g. rsync with timestamp comparisons.