Rewrite _idmap to make fewer assumptions
Previously, _idmap made horrible assumptions about the [UG]IDs of
certain entities being adjacent or within a few integers of each other,
as well as the username of the process and so on.

In exchange, in order to have > 5 lines in the resulting {uid,gid}_map
files, this requires kernel >= 4.15.