af-rm-container: fix nasty strn?cmp bug
Applying sizeof to a member of an array of strings (char *[]) is a no-no
- the string will decay to a pointer, so sizeof will just return the
size of the pointer (on x86_64, 8). This happened to work mostly by
accident since the excluded paths list happened to be unique at 8
characters. It is not a huge deal anyway since the paths it was breaking
should be unmounted anyway (except for / and /af/libexec, but "/af/libe"
is unique enough that it still skipped it correctly, and / is short
enough).

Instead, do some magic with macros so that the size of the strings are
recorded as constants, and while we're at it, merge m_dirs and
m_contents into one thing.

This fix brought to you by -Wsizeof-pointer-memaccess.